feat: 회원가입 시 링크 저장 기능 추가, metadata를 위한 컴포넌트 구조 변경
