+++ src/app/(public)/signup/link/page.tsx
@@ -1,115 +1,19 @@
-"use client";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { ChangeEvent, useState } from "react";
+import { Metadata } from "next";
 
-import toast from "react-hot-toast";
-
-import { LinkAddButtons } from "@/components/signup/link-add-buttons";
-import { LinkListItem } from "@/components/signup/link-list-item";
+import { LinkListEditor } from "@/components/signup/link-list-editor";
 import {
   AppHeader,
   AppHeaderCenter,
   AppHeaderLeft,
   AppHeaderRight,
 } from "@/components/ui/app-header";
-import { Button } from "@/components/ui/button";
 import { Heading } from "@/components/ui/heading";
-import { ENV } from "@/constants/env";
-import { getSnsUrl } from "@/lib/utils";
-import { LinkType } from "@/types/link";
-/* import { Metadata } from "next";
 
-export const metadata: Metadata = {
-  title: "회원가입 | 링꾸",
-}; */
-
-export interface AddLinkInputProps {
-  /** 링크 id */
-  id: number;
-  /** 링크 URL */
-  url?: string;
-  /** 링크 type*/
-  type?: LinkType;
-}
-
-const iconLists = [
-  { type: "custom", iconLabel: "커스텀" },
-  {
-    type: "instagram",
-    iconLabel: "인스타그램",
-  },
-  {
-    type: "facebook",
-    iconLabel: "페이스북",
-  },
-
-  {
-    type: "threads",
-    iconLabel: "쓰레드",
-  },
-];
-
-/**링크저장 API --> 다음 버튼을 누르면 모두 한번에 저장됨 */
-async function addLinks(id: number, linkInputs: AddLinkInputProps[]) {
-  console.log("insert link array");
-
-  const data = JSON.stringify({ links: linkInputs });
-
-  try {
-    const response = await fetch(`${ENV.apiUrl}/api/links?user_id=${id}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: data,
-    });
-
-    if (!response.ok) {
-      throw new Error("URL 등록에 실패했습니다");
-    }
-  } catch (error) {
-    console.log("링크 등록 중 오류 발생: ", error);
-  }
-}
+const metadata: Metadata = {
+  title: "회원가입",
+};
 
 export default function RegisterLinksPage() {
-  const [linkInputs, setLinkInputs] = useState<AddLinkInputProps[]>([]);
-  const router = useRouter();
-
-  const handleNext = async () => {
-    try {
-      await addLinks(1, linkInputs);
-      router.push("/signup/welcome");
-    } catch (error) {
-      console.log("링크 저장 중 오류 발생: ", error);
-      toast("링크 추가에 실패했어요. 잠시후에 다시 시도해주세요");
-    }
-  };
-
-  const handleAddLink = (type: LinkType) => {
-    const newLink = {
-      id: Date.now(),
-      url: getSnsUrl(type),
-      type,
-    };
-
-    setLinkInputs([...linkInputs, newLink]);
-  };
-
-  const handleChangeUrl = (id: number, e: ChangeEvent<HTMLInputElement>) => {
-    const wholeUrl = e.target.value;
-
-    setLinkInputs((prev) =>
-      prev.map((link) => (link.id === id ? { ...link, url: wholeUrl } : link)),
-    );
-    console.log(linkInputs);
-  };
-
-  const onDelete = (id: number) => {
-    setLinkInputs(linkInputs.filter((input) => input.id !== id));
-  };
-
   return (
     <>
       <AppHeader>
@@ -123,28 +27,7 @@ export default function RegisterLinksPage() {
           <div className="px-3 pb-12">
             <Heading variant="subtitle1">사용하실 웹사이트 링크를 연결해보세요</Heading>
           </div>
-          <LinkAddButtons iconLists={iconLists} onAdd={handleAddLink} />
-          <section className="w-full">
-            <ul className="mt-4 flex flex-col gap-4 px-2">
-              {linkInputs.map((item) => (
-                <LinkListItem
-                  key={item.id}
-                  type={item.type}
-                  {...item}
-                  onDelete={onDelete}
-                  onChangeUrl={handleChangeUrl}
-                />
-              ))}
-            </ul>
-            <div className="mt-20 flex flex-col gap-2 px-2">
-              <Button variant="primary" size="large" onClick={handleNext}>
-                다음
-              </Button>
-              <Button variant="text" size="large">
-                <Link href="/signup/welcome">건너뛰기</Link>
-              </Button>
-            </div>
-          </section>
+          <LinkListEditor />
         </div>
       </main>
     </>
